Most of whats needed for deliverable 6, no admin signin.
The code is very sloppy, I will need to go back and straighten it out as well as document. As of now you can change the payment options expiration dates but that is all, there are no database methods to change anything else but that could be made quickly. Instead of being able to create payment methods up to 4, there are 4 that you can edit/create by selecting a slot 1-4. Logging in and out works, as well as changing the necessary parts of the profile. Because I could not figure out how to carry the accountID along with the session value, at the top of the editPage there is a place where you need to enter email and password in order to view user information or change any info. Because of this system your password is required in order to update your password.